+++ src/components/notice/index.tsx
@@ -1,5 +1,3 @@
-export { default as HeaderWithNav } from "./header/HeaderWithNav";
-
 export { default as Header } from "./header/Header";
 
 export { default as TimeInformation } from "./timeInfro/TimeInformation";
